rename all_demos.dart to all.dart
