+++ examples/material_gallery/lib/gallery/app.dart
@@ -5,7 +5,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart' show timeDilation;
 
-import '../demo/all_demos.dart';
+import '../demo/all.dart';
 import 'home.dart';
 
 final Map<String, WidgetBuilder> kRoutes = <String, WidgetBuilder>{
