+++ examples/material_gallery/lib/gallery/home.dart
@@ -5,7 +5,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-import '../demo/all_demos.dart';
+import '../demo/all.dart';
 import 'drawer.dart';
 import 'header.dart';
 import 'item.dart';
